requirements.txt
pandas==1.5.3
ecg_plot==0.2.8
tqdm==4.65.0
numpy==1.24.2
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.13.1
scikit_learn==1.2.2
hydra_core==1.3.2
omegaconf==2.3.0
timm
fastervit
datasets
transformers
peft
timm
tabulate